update permission edit

+++ src/Controllers/CategoryController.php
@@ -226,7 +226,7 @@ class CategoryController extends Controller
                 }
                 $nestedData['No'] = $DT_RowIndex;
 
-                if (checkPermission('is_admin')) {
+                if (config('MasterCrudConfig.MASTER_DIRECT_EDIT') == true && checkPermission('is_admin')) {
                     $btn .= '<a href="' . route('master.' . $sheet_slug . '.edit', $row->No) . '" class="btn btn-primary btn-sm">Update</a> ';
                     $btn .= '<a href="' . route('master.' . $sheet_slug . '.destroy', $row->No) . '" onclick="notificationBeforeDelete(event,this)" class="btn btn-danger btn-sm">Delete</a>';
                 } else {
